toggle theme throughout
