+++ Client/src/app/app.component.html
@@ -6,11 +6,6 @@
       </a>
     </h1>
     <div class="filler"></div>
-    <div class="toggle">
-      <mat-slide-toggle [(ngModel)]="ts.isDarkTheme" [color]="toggleColor" (change)="ts.onToggle($event)">
-        dark theme
-      </mat-slide-toggle>
-    </div>
     <a class="badge" href="https://www.nuget.org/packages/Skender.Stock.Indicators"
       matTooltip="NuGet package for Stock Indicators for .NET">
       <img src="https://img.shields.io/nuget/v/skender.stock.indicators?color=black&logo=NuGet&label=NuGet" />
