+++ Client/src/app/app.component.scss
@@ -25,7 +25,3 @@ mat-toolbar {
   margin-top: 0.25em;
   margin-right: 16px;
 }
-
-.toggle {
-  margin-right: 32px;
-}
