+++ Client/src/app/chart/chart.component.html
@@ -16,8 +16,8 @@
 </div>
 
 <!-- ADD INDICATOR -->
-<button mat-fab color="accent" class="button-listing" (click)="openPickList()" matTooltip="add a new indicator">
-  <mat-icon>add</mat-icon>/<mat-icon>remove</mat-icon>
+<button mat-mini-fab color="accent" class="button-listing" (click)="openPickList()" matTooltip="add a new indicator">
+  <mat-icon>settings</mat-icon>
 </button>
 
 <!-- CHART (OSCILLATORS) -->
