+++ Client/src/app/chart/chart.component.scss
@@ -5,8 +5,12 @@
   margin: 0;
   border-style: none;
   border-width: 0;
-  min-height: 70vh;
+  min-height: 55vh;
   max-height: calc(100vh - 64px);
+
+  @media (max-height: 400px){
+    height: calc(100vh - 64px); // full-screen (with toolbar)
+  }
 }
 
 .chart-oscillator-container {
@@ -29,4 +33,8 @@
     color: $color-primary-500 !important;
     background-color: $color-accent-800 !important;
   }
+}
+
+.toggle {
+  margin-left: 16px;
 }
