+++ Client/src/app/chart/chart.service.ts
@@ -1,6 +1,10 @@
 import { Injectable } from '@angular/core';
-import { ApiService } from './api.service';
+import { HttpErrorResponse } from '@angular/common/http';
 
+import { ApiService } from './api.service';
+import { StyleService } from '../style.service';
+
+import Chart from 'chart.js/auto';  // import all default options
 import 'chartjs-adapter-date-fns';
 import 'chartjs-chart-financial';
 
@@ -9,9 +13,9 @@ import { add, parseISO } from 'date-fns';
 import { Guid } from "guid-typescript";
 
 import {
-  Chart,
   ChartConfiguration,
   ChartDataset,
+  FinancialDataPoint,
   FontSpec,
   ScaleOptions,
   ScatterDataPoint,
@@ -38,9 +42,9 @@ import {
   IndicatorParamConfig,
   IndicatorResult,
   IndicatorResultConfig,
-  IndicatorSelection
+  IndicatorSelection,
+  Quote
 } from './chart.models';
-import { HttpErrorResponse } from '@angular/common/http';
 
 Chart.register(
   CandlestickController,
@@ -53,7 +57,8 @@ Chart.register(
 export class ChartService {
 
   constructor(
-    private readonly api: ApiService
+    private readonly api: ApiService,
+    private readonly ts: StyleService
   ) { }
 
   yAxisTicks: Tick[] = [];
@@ -65,7 +70,7 @@ export class ChartService {
   baseConfig() {
 
     const commonXaxes = this.commonXAxes();
-
+    const gridColor = this.ts.isDarkTheme ? '#424242' : '#CCCCCC';
     const backgroundPlugin =
     {
       id: 'background',
@@ -73,7 +78,7 @@ export class ChartService {
         const ctx = chart.canvas.getContext('2d');
         ctx.save();
         ctx.globalCompositeOperation = 'destination-over';
-        ctx.fillStyle = '#212121';
+        ctx.fillStyle = this.ts.isDarkTheme ? '#212121' : 'white';
         ctx.fillRect(0, 0, chart.width, chart.height);
         ctx.restore();
       }
@@ -97,7 +102,7 @@ export class ChartService {
             display: false
           },
           tooltip: {
-            enabled: true,
+            enabled: false,
             mode: 'index',
             intersect: false
           },
@@ -113,7 +118,7 @@ export class ChartService {
         },
         responsive: true,
         maintainAspectRatio: false,
-
+        animation: false,
         scales: {
           xAxis: commonXaxes,
           yAxis: {
@@ -131,7 +136,7 @@ export class ChartService {
                 lineHeight: 1
               },
               showLabelBackdrop: true,
-              backdropColor: '#212121',
+              backdropColor: this.ts.isDarkTheme ? '#212121' : 'white',
               backdropPadding: {
                 top: 0,
                 left: 2,
@@ -148,7 +153,7 @@ export class ChartService {
                 if (context.tick.label === null) {
                   return 'transparent';
                 } else {
-                  return '#424242';
+                  return gridColor;
                 }
               },
 
@@ -307,7 +312,6 @@ export class ChartService {
 
     // add to collection
     this.selections.push(selection);
-    // TODO: save to cache
 
     // lookup config data
     const listing = this.listings.find(x => x.uiid == selection.uiid);
@@ -361,7 +365,7 @@ export class ChartService {
     const selections = this.selections;
 
     // remove unsavable data
-    selections.forEach((selection:IndicatorSelection)=>{
+    selections.forEach((selection: IndicatorSelection) => {
       selection.chart = undefined;
     });
 
@@ -437,15 +441,26 @@ export class ChartService {
       chartConfig.data.datasets.push(r.dataset);
     });
 
-    // compose chart
+
+    // compose html
+    const body = document.getElementById("main-content");
+    const containerId = `${selection.ucid}-container`;
+
+    // pre-delete, if exists (needed for theme change)
+    const existing = document.getElementById(containerId);
+    if (existing != null) {
+      body.removeChild(existing);
+    }
+
+    // create chart container
+    const container = document.createElement('div') as HTMLDivElement;
+    container.id = containerId
+    container.className = "chart-oscillator-container";
+
+    // add chart
     const myCanvas = document.createElement('canvas') as HTMLCanvasElement;
     myCanvas.id = selection.ucid;
-
-    const container = document.createElement('div') as HTMLDivElement;
-    container.id = `${selection.ucid}-container`;
-    container.className = "chart-oscillator-container";
     container.appendChild(myCanvas);
-    const body = document.getElementById("main-content");
     body.appendChild(container);
 
     if (selection.chart) selection.chart.destroy();
@@ -501,7 +516,7 @@ export class ChartService {
       content: [label],
       font: legendFont,
       color: fontColor,
-      backgroundColor: 'rgba(33,33,33,0.5)',
+      backgroundColor: this.ts.isDarkTheme ? 'rgba(33,33,33,0.5)' : 'rgba(255,255,255,0.9)',
       padding: 0,
       position: 'start',
       xScaleID: 'xAxis',
@@ -515,6 +530,187 @@ export class ChartService {
     return annotation;
   }
 
+  // DATA OPERATIONS
+  loadCharts() {
+    this.api.getQuotes()
+      .subscribe({
+        next: (q: Quote[]) => {
+
+          this.loadOverlayChart(q);
+
+          // load default selections
+          this.api.getListings()
+            .subscribe({
+              next: (listings: IndicatorListing[]) => {
+                this.listings = listings;
+                this.loadSelections()
+              },
+              error: (e: HttpErrorResponse) => { console.log(e); }
+            });
+        },
+        error: (e: HttpErrorResponse) => { console.log(e); }
+      });
+  }
+
+  loadOverlayChart(quotes: Quote[]) {
+
+    const chartConfig = this.baseOverlayConfig();
+    const candleOptions = Chart.defaults.elements["candlestick"];
+
+    // custom border colors
+    candleOptions.color.up = '#1B5E20';
+    candleOptions.color.down = '#B71C1C';
+    candleOptions.color.unchanged = '#616161';
+
+    candleOptions.borderColor = {
+      up: candleOptions.color.up,
+      down: candleOptions.color.down,
+      unchanged: candleOptions.color.unchanged
+    };
+
+    const price: FinancialDataPoint[] = [];
+    const volume: ScatterDataPoint[] = [];
+    const barColor: string[] = [];
+
+    let sumVol = 0;
+
+    quotes.forEach((q: Quote) => {
+
+      price.push({
+        x: new Date(q.date).valueOf(),
+        o: q.open,
+        h: q.high,
+        l: q.low,
+        c: q.close
+      });
+
+      volume.push({
+        x: new Date(q.date).valueOf(),
+        y: q.volume
+      });
+      sumVol += q.volume;
+
+      const c = (q.close >= q.open) ? '#1B5E2060' : '#B71C1C60';
+      barColor.push(c);
+    });
+
+    // define base datasets
+    chartConfig.data = {
+      datasets: [
+        {
+          type: 'candlestick',
+          label: 'Price',
+          data: price,
+          yAxisID: 'yAxis',
+          borderColor: candleOptions.borderColor,
+          order: 75
+        },
+        {
+          type: 'bar',
+          label: 'Volume',
+          data: volume,
+          yAxisID: 'volumeAxis',
+          backgroundColor: barColor,
+          borderWidth: 0,
+          order: 76
+        }
+      ]
+    };
+
+    // get size for volume axis
+    const volumeAxisSize = 20 * (sumVol / volume.length) || 0;
+    chartConfig.options.scales.volumeAxis.max = volumeAxisSize;
+
+    // compose chart
+    if (this.chartOverlay) this.chartOverlay.destroy();
+    const myCanvas = document.getElementById("chartOverlay") as HTMLCanvasElement;
+    this.chartOverlay = new Chart(myCanvas.getContext('2d'), chartConfig);
+  }
+
+  loadSelections() {
+
+    // get from cache
+    const selections = JSON.parse(localStorage.getItem('selections'));
+
+    if (selections) {
+      selections.forEach((selection: IndicatorSelection) => {
+        this.addSelection(selection);
+      });
+    }
+    else { // add defaults
+      const def1 = this.defaultSelection("EMA");
+      this.addSelection(def1);
+
+      const def2 = this.defaultSelection("BB");
+      this.addSelection(def2);
+
+      const def3 = this.defaultSelection("STO");
+      this.addSelection(def3);
+
+      const def4 = this.defaultSelection("RSI");
+      def4.params.find(x => x.paramName == "lookbackPeriods").value = 5;
+      this.addSelection(def4);
+    }
+  }
+
+  updateData() {
+
+    // update selections data
+    this.selections.forEach((selection: IndicatorSelection) => {
+
+      // lookup config data
+      const listing = this.listings.find(x => x.uiid == selection.uiid);
+
+      this.api.getSelectionData(selection, listing)
+        .subscribe({
+          next: () => {
+            if (listing.chartType != 'overlay') {
+              selection.chart.update();
+            };
+          },
+          error: (e: HttpErrorResponse) => { console.log(e); }
+        });
+    });
+
+    // update primary data
+    this.api.getQuotes()
+      .subscribe({
+        next: (quotes: Quote[]) => {
+
+          // TODO: refactor redundant code (see load)
+          const price: FinancialDataPoint[] = [];
+          const volume: ScatterDataPoint[] = [];
+          let sumVol = 0;
+
+          quotes.forEach((q: Quote) => {
+            price.push({
+              x: new Date(q.date).valueOf(),
+              o: q.open,
+              h: q.high,
+              l: q.low,
+              c: q.close
+            });
+            volume.push({
+              x: new Date(q.date).valueOf(),
+              y: q.volume
+            });
+            sumVol += q.volume;
+
+            // get size for volume axis
+            const volumeAxisSize = 20 * (sumVol / volume.length) || 0;
+            this.chartOverlay.options.scales.volumeAxis.max = volumeAxisSize;
+            this.chartOverlay.update();
+          });
+        },
+        error: (e: HttpErrorResponse) => { console.log(e); }
+      });
+  }
+
+  resetChartTheme() {
+
+    this.selections = [];
+    this.loadCharts();
+  }
 
   // helper functions
   getGuid(prefix: string = "chart"): string {
